Positive anything is better than negative nothing

+++ 0x01-variables_if_else_while/0-positive_or_negative.c
@@ -12,8 +12,7 @@ int main(void)
 	int n;
 
 	srand(time(0));
-	n = rand() - (int)((double)RAND_MAX / 2.0);
-
+	n = rand() - RAND_MAX / 2.0;
 	if (n > 0)
 	{
 		printf("%d is positive\n", n);
